+++ AffirmSDKTests/AffirmSDKTests.m
@@ -7,10 +7,10 @@
 //
 
 #import <XCTest/XCTest.h>
-#import "AffirmConfiguration.h"
-#import "AffirmUtils.h"
-#import "AffirmRequest.h"
-#import "AffirmCardValidator.h"
+#import "../AffirmSDK/AffirmConfiguration.h"
+#import "../AffirmSDK/AffirmUtils.h"
+#import "../AffirmSDK/AffirmRequest.h"
+#import "../AffirmSDK/AffirmCardValidator.h"
 
 @interface AffirmSDKTests : XCTestCase